+++ fedcore/models/network_impl/base_nn_model.py
@@ -1,7 +1,7 @@
 import os
 from copy import deepcopy
 from datetime import datetime
-from functools import reduce
+from functools import reduce, partial
 from operator import iadd
 from pathlib import Path
 from typing import Callable, Literal, Optional
@@ -19,7 +19,7 @@ from tqdm import tqdm
 from fedcore.api.utils.data import DataLoaderHandler
 from fedcore.data.data import CompressionInputData
 from fedcore.losses.utils import _get_loss_metric
-from fedcore.repository.constanst_repository import default_device
+from fedcore.repository.constanst_repository import default_device, Hooks
 from fedcore.architecture.abstraction.accessor import Accessor
 
 
@@ -59,6 +59,7 @@ class BaseNeuralModel:
         )  # loss which evaluates model structure
         self.enforced_training_loss = self.params.get("enforced_training_loss", None)
         self.device = self.params.get('device', default_device())
+        self._optimizer_gen = partial(torch.optim.Adam, lr=self.learning_rate)
 
         self.is_operation = self.params.get('is_operation', False) ###
         self.save_each = self.params.get('save_each', None)
@@ -74,6 +75,10 @@ class BaseNeuralModel:
         self.model = None
         self.target = None
         self.task_type = None
+
+        # add hooks
+        self._on_epoch_end = []
+        self._on_epoch_start = []
 
     def __check_and_substitute_loss(self, train_data: InputData):
         if (
@@ -105,10 +110,9 @@ class BaseNeuralModel:
         self.__check_and_substitute_loss(input_data)
         if self.model is None:
             self.model = input_data.target
+        self._init_hooks()
         self.optimised_model = self.model
-        self.optimizer = torch.optim.Adam(
-            self.model.parameters(), lr=self.learning_rate
-        )
+        self.optimizer = self._optimizer_gen(self.model.parameters())
         self.model.to(self.device)
 
         fit_output = Either(
@@ -197,19 +201,21 @@ class BaseNeuralModel:
             if epoch > 2:
                 # Freeze batch norm mean and variance estimates
                 self.model.apply(torch.nn.intrinsic.qat.freeze_bn_stats)
-            if self._check_saving(epoch):
-                self.save_model(epoch, self.name)
+    #         if self._check_saving(epoch):
+    #             self.save_model(epoch, self.name)
 
-    def _check_saving(self, epoch) -> bool:
-        if not self.save_each:
-            return False
-        if self.save_each != -1:
-            return not epoch % self.save_each
-        else:
-            return epoch == self.epochs
+    # def _check_saving(self, epoch) -> bool:
+    #     if not self.save_each:
+    #         return False
+    #     if self.save_each != -1:
+    #         return not epoch % self.save_each
+    #     else:
+    #         return epoch == self.epochs
 
     def _default_train(self, train_loader, model, custom_loss: dict = None, val_loader=None):
         for epoch in range(1, self.epochs + 1):
+            for hook in self._on_epoch_start:
+                hook(epoch=epoch)
             self.model.train()
             model_loss, avg_loss = self._train_loop(train_loader, model, custom_loss)
             if model_loss is not None:
@@ -220,36 +226,40 @@ class BaseNeuralModel:
                 )
             else:
                 print("Epoch: {}, Average loss {}".format(epoch, avg_loss))
-            if epoch % self.eval_each == 0 and val_loader is not None:
-                print('Model Validation:' , self._eval(self.model, val_loader, custom_loss))
-            if self._check_saving(epoch):
-                self.save_model(epoch, self.name)            
+
+            for hook in self._on_epoch_end:
+                hook(epoch=epoch, val_loader=val_loader, custom_loss=custom_loss)
+            
+            # if epoch % self.eval_each == 0 and val_loader is not None:
+            #     print('Model Validation:' , self._eval(self.model, val_loader, custom_loss))
+            # if self._check_saving(epoch):
+            #     self.save_model(epoch, self.name)            
     
-    def save_model(self, epoch, name=''):
-        name = name or self.params.get('name', '')
-        path_pref = Path(self.checkpoint_folder)
-        save_only = self.params.get('save_only', '')
-        to_save = self.model if not save_only else Accessor.get_module(self.model, save_only)
-        try:
-            path = path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}.pth")
-            torch.save(
-                to_save,
-                path,
-            )
-        except Exception as x:
-            if os.path.exists(path):
-                os.remove(path)
-            print('Basic saving failed. Trying to use jit. \nReason: ', x.args[0])
-            try:
-                path = path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}_jit.pth")
-                torch.jit.save(torch.jit.script(to_save), path)
-            except Exception as x: 
-                if os.path.exists(path):
-                    os.remove(path)
-                print('JIT saving failed. saving weights only. \nReason: ', x.args[0])
-                torch.save(to_save.state_dict(), 
-                            path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}_state.pth")
-                )        
+    # def save_model(self, epoch, name=''):
+    #     name = name or self.params.get('name', '')
+    #     path_pref = Path(self.checkpoint_folder)
+    #     save_only = self.params.get('save_only', '')
+    #     to_save = self.model if not save_only else Accessor.get_module(self.model, save_only)
+    #     try:
+    #         path = path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}.pth")
+    #         torch.save(
+    #             to_save,
+    #             path,
+    #         )
+    #     except Exception as x:
+    #         if os.path.exists(path):
+    #             os.remove(path)
+    #         print('Basic saving failed. Trying to use jit. \nReason: ', x.args[0])
+    #         try:
+    #             path = path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}_jit.pth")
+    #             torch.jit.save(torch.jit.script(to_save), path)
+    #         except Exception as x: 
+    #             if os.path.exists(path):
+    #                 os.remove(path)
+    #             print('JIT saving failed. saving weights only. \nReason: ', x.args[0])
+    #             torch.save(to_save.state_dict(), 
+    #                         path_pref.joinpath(f"model_{name}{now_for_file()}_{epoch}_state.pth")
+    #             )        
 
     def predict(self, input_data: InputData, output_mode: str = "default"):
         """
@@ -328,3 +338,14 @@ class BaseNeuralModel:
     @property
     def is_quantised(self):
         return getattr(self, '_is_quantised', False)
+    
+    def _init_hooks(self):
+        for hook_elem in Hooks:
+            hook = hook_elem.value
+            if not self.params.get(hook._SUMMON_KEY, None):
+                continue
+            hook = hook(self.params, self.model)
+            if hook._hook_place == 'post':
+                self._on_epoch_end.append(hook)
+            else:
+                self._on_epoch_start.append(hook)
